fixes Android halting events, loosens regex for auth token
Overhauls the execution of InAppBrowser. Instead of calling the exec
method, this creates an InAppBrowser instance and creates a handle to
add listeners on.

Android implementation of the InAppBrowser is a clusterfuck and will
halt
all events until the user taps the done button or the IAB is manually
closed.
However if you hide the IAB and call `.show`, it will yield and let the
captured
events flow in the meteor window. If on Android we keep calling show
until
an error or auth code is sniffed from the IAB loadstop event.